Move network::auth2 to network::auth
Since auth2 has replace auth completely, we can now rename auth2 to
auth, as was the original intent.
