+++ libsplinter/src/admin/service/shared.rs
@@ -1901,7 +1901,7 @@ mod tests {
     use crate::circuit::directory::CircuitDirectory;
     use crate::keys::insecure::AllowAllKeyPermissionManager;
     use crate::mesh::{Envelope, Mesh};
-    use crate::network::auth2::AuthorizationManager;
+    use crate::network::auth::AuthorizationManager;
     use crate::network::connection_manager::authorizers::{Authorizers, InprocAuthorizer};
     use crate::network::connection_manager::ConnectionManager;
     use crate::network::peer_manager::{PeerManager, PeerManagerConnector};
