+++ libsplinter/src/network/auth/connection_manager.rs
