+++ libsplinter/src/network/auth/pool.rs
